// project/h5/src/components/OpenDatePicker/types.ts
export interface OpenDatePickerOptions {
  /** 相同id弹窗只会出现一个 */
  id?: string
  value: any
  field: string
  props?: any
  type?: 1 | 2 | 3
  formatData?: (d: any) => any
  /** 关闭回调 */
  onClose?: () => void
  onConfirm?: (v: string) => void
  onReady?: (v: any) => void
  onSearch?: (v: string) => void
}

export interface OpenDatePickerOptionsStyle {
  width?: number
  height?: number
  top?: number
  left?: number
}